refactor: package the project
