+++ docs/README.md
@@ -79,3 +79,9 @@
 * `Game` - 게임에 관한 부분을 담당하는 객체
 * `Car` - 자동차 객체
 * `ExceptionHandler` - 예외 처리를 담당하는 객체
+
+# 추가 객체
+
+---
+* `CarStorage` - 차의 정보들을 저장하고 읽고 갱신할 수 있는 객체
+* `GameProgress` - 횟수 마다의 진행 상황을 나타낼 수 있는 객체
