+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.commander.GameHandler;
+import racingcar.instance.InstanceContainer;
+
 public class Application {
     public static void main(String[] args) {
         InstanceContainer container = new InstanceContainer();
